# internal/api/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "PlaylistLabs User Management API",
    "description": "REST API for PlaylistLabs User Management open-source panel.\nAllows third-party systems, and automation tools to manage **Playlists & Sync**, **Managed Users**, **Team Members / Collaborators**, and **API Tokens**.\n\n### Authentication\nAuthentication is performed strictly using an internal API Token (`plt_live_...`) in the canonical **`X-API-Token`** header:\n- Header: `X-API-Token: <your_token>`\n\n### Optional Token Password\nIf the API Token was configured with an optional password, include it in the **`X-Token-Password`** header:\n- Header: `X-Token-Password: <your_password>`\n\n### Anti-Brute-Force & IP Banning\nInvalid tokens or repeated incorrect token passwords trigger automatic IP rate tracking.\nExceeding the maximum allowed failed attempts within the security window causes an **immediate IP ban** (`403 Forbidden`). Once banned, subsequent requests from the IP are rejected with no further logging.\n",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/api",
      "description": "Current server base API path"
    }
  ],
  "security": [
    {
      "ApiTokenAuth": []
    },
    {
      "ApiTokenAuth": [],
      "TokenPasswordAuth": []
    }
  ],
  "paths": {
...
    "/team": {
      "get": {
        "summary": "List team members",
        "description": "Retrieves the team hierarchy (administrators and collaborators). Scoped to the caller's allowed management branch.",
        "operationId": "getTeamMembers",
        "tags": [
          "Team Management"
        ],
        "responses": {
          "200": {
            "description": "Team members list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TeamMember"
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      },
      "post": {
        "summary": "Create a team member (Admin or Collaborator)",
        "description": "Adds a new staff member. Requires team creation permissions (`can_create_admins` or `can_create_collaborators`).",
        "operationId": "createTeamMember",
        "tags": [
          "Team Management"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTeamMemberRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Team member created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TeamMember"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
    "/team/{id}": {
      "get": {
        "summary": "Get team member details",
        "description": "Retrieves details of a specific team member.",
        "operationId": "getTeamMember",
        "tags": [
          "Team Management"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "example": 2
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Team member details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TeamMember"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "put": {
        "summary": "Update a team member",
        "description": "Updates roles, playlist permissions, and access privileges. Revoking `can_manage_api_tokens` automatically cascades down to all subordinate collaborators.",
        "operationId": "updateTeamMember",
        "tags": [
          "Team Management"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "example": 2
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTeamMemberRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Team member updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TeamMember"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "delete": {
        "summary": "Delete a team member",
        "description": "Deletes a collaborator or administrator account. Protected against self-deletion and last-admin deletion.",
        "operationId": "deleteTeamMember",
        "tags": [
          "Team Management"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "example": 2
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Member deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Team member deleted successfully"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "ApiTokenAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Token",
        "description": "PlaylistLabs internal API Token (format: plt_live_...)"
      },
      "TokenPasswordAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Token-Password",
        "description": "Optional password configured during API token generation (leave empty if no password was set)"
      }
    },
    "schemas": {
...
      "TeamMember": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 2
          },
          "username": {
            "type": "string",
            "example": "collaborator_john"
          },
          "role": {
            "type": "string",
            "enum": [
              "admin",
              "collaborator"
            ],
            "example": "collaborator"
          },
          "manage_all_playlists": {
            "type": "boolean",
            "example": false
          },
          "allowed_playlist_ids": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "example": [
              "1001",
              "1002"
            ]
          },
          "can_see_all_users": {
            "type": "boolean",
            "example": false
          },
          "can_create_collaborators": {
            "type": "boolean",
            "example": true
          },
          "can_create_admins": {
            "type": "boolean",
            "example": false
          },
          "can_manage_api_tokens": {
            "type": "boolean",
            "example": true
          },
          "created_by": {
            "type": "integer",
            "nullable": true,
            "example": 1
          },
          "created_by_username": {
            "type": "string",
            "nullable": true,
            "example": "admin"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "CreateTeamMemberRequest": {
        "type": "object",
        "required": [
          "username",
          "password"
        ],
        "properties": {
          "username": {
            "type": "string",
            "example": "new_staff"
          },
          "password": {
            "type": "string",
            "example": "StaffSecure123!"
          },
          "role": {
            "type": "string",
            "enum": [
              "admin",
              "collaborator"
            ],
            "default": "collaborator"
          },
          "manage_all_playlists": {
            "type": "boolean",
            "default": false
          },
          "allowed_playlist_ids": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int64"
            },
            "example": [
              1001
            ]
          },
          "can_see_all_users": {
            "type": "boolean",
            "default": false
          },
          "can_create_collaborators": {
            "type": "boolean",
            "default": false
          },
          "can_create_admins": {
            "type": "boolean",
            "default": false
          },
          "can_manage_api_tokens": {
            "type": "boolean",
            "default": false
          }
        }
      },
      "UpdateTeamMemberRequest": {
        "type": "object",
        "properties": {
          "password": {
            "type": "string",
            "description": "Leave empty to retain current password"
          },
          "role": {
            "type": "string",
            "enum": [
              "admin",
              "collaborator"
            ]
          },
          "manage_all_playlists": {
            "type": "boolean"
          },
          "allowed_playlist_ids": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int64"
            }
          },
          "can_see_all_users": {
            "type": "boolean"
          },
          "can_create_collaborators": {
            "type": "boolean"
          },
          "can_create_admins": {
            "type": "boolean"
          },
          "can_manage_api_tokens": {
            "type": "boolean"
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "example": "invalid_request"
          },
          "message": {
            "type": "string",
            "example": "Invalid parameters supplied"
          }
        }
      }
    },
    "responses": {
      "BadRequest": {
        "description": "Invalid request parameters or body",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Authentication failed, missing token, expired token, or invalid token password",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "Forbidden": {
        "description": "Access denied, insufficient permissions, API tokens disabled for this account, or IP address blocked",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "NotFound": {
        "description": "Requested resource was not found",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      }
    }
  }
}